% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseq.deg.plot.R
\name{arseq.deg.plot}
\alias{arseq.deg.plot}
\title{Heatmap of the differentially expressed genes}
\usage{
arseq.deg.plot(deg, data, dds, save.plot = FALSE, save.dir = getwd())
}
\arguments{
\item{deg}{Differtially expressed genes calculated using DESeq2}

\item{data}{Normalized count's matrix}

\item{dds}{DeSes2 object}

\item{save.plot}{Logical. Argument to state if the plot needs to be saved in disk. Default: FALSE}

\item{save.dir}{Location to save the plot when 'save.plot=TRUE'. Default: Working Directory.}
}
\value{
Heatmap of the differentially expressed genes.
}
\description{
Generating a heatmap of the differentially expressed genes
}
