% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseq.ensembl2genename.R
\name{arseq.ensembl2genename}
\alias{arseq.ensembl2genename}
\title{Gene ID converter}
\usage{
arseq.ensembl2genename(data, ensemblmirror, species)
}
\arguments{
\item{data}{Expression matrix with ensemble id's as the first column}

\item{species}{Only applies to converting ENSEMBL IDs to Gene names (not to enrichment analysis).  Species you want to use. To see the different datasets available you can use do: library(biomaRt); followed by mart = useEnsembl('ENSEMBL_MART_ENSEMBL'); followed by listDatasets(mart). Default: 'hsapiens_gene_ensembl'.}
}
\value{
Expression matrix with gene names as the first column
}
\description{
Convert ensembl id's to hugo gene names. The function also reduces multiple transcripts into a single gene by taking the sum of all transcripts.
}
\examples{
\dontrun{
data <- arseq.ensembl2genename (example_data)
}
}
