% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseq.euclid.dist.plot.R
\name{arseq.euclid.dist.plot}
\alias{arseq.euclid.dist.plot}
\title{Euclidean distance heatmap}
\usage{
arseq.euclid.dist.plot(euclid.dist)
}
\arguments{
\item{euclid.dist}{Euclidean distance matrix.}
}
\value{
Euclidean distance heatmap.
}
\description{
Plots a heatmap of the distance between samples using pheatmap.
}
\examples{
\dontrun{
euclid.dist <- arseq.euclid.dist (example_dds)
euclid.plot <- arseq.euclid.dist.plot (euclid.dist)
}
}
