% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseq.euclid.dist.R
\name{arseq.euclid.dist}
\alias{arseq.euclid.dist}
\title{Euclidean distance}
\usage{
arseq.euclid.dist(dds)
}
\arguments{
\item{dds}{DESeq2 object}
}
\value{
Euclidean distance matrix
}
\description{
Calculate the Euclidean distance between samples.
}
\examples{
euclid.dist <- arseq.euclid.dist (example_dds)
}
