% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseq.go.enrich.plot.R
\name{arseq.go.enrich.plot}
\alias{arseq.go.enrich.plot}
\title{Gene Ontology (GO) term enrichment}
\usage{
arseq.go.enrich.plot(go.enrich)
}
\arguments{
\item{go.enrich}{Enrichment dataframe. Output of the function 'arseq.go.enrich'}
}
\value{
GO enrichment plot
}
\description{
Plotting the top 5 enriched term from each category
}
\examples{
\dontrun{
go.enrich <- arseq.go.enrich (example_deg)
go.plot <- arseq.go.enrich.plot (go.enrich)
}
}
