% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseq.go.enrich.R
\name{arseq.go.enrich}
\alias{arseq.go.enrich}
\title{Gene Ontology (GO) term enrichment}
\usage{
arseq.go.enrich(deg, Padj = 0.05)
}
\arguments{
\item{deg}{Differentially expressed genes dataframe returned by DESeq2 analysis}

\item{Padj}{Numeric. Adjusted P Value cut off to define the differentially expressed genes. Default = 0.05}
}
\value{
Enrichment dataframe
}
\description{
Performing GO term enrichment analysis on the significantly differentially expressed genes.
}
\examples{
\dontrun{
go.enrich <- arseq.go.enrich (example_deg)
}
}
