% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseq.gsea.preprocess.R
\name{arseq.gsea.preprocess}
\alias{arseq.gsea.preprocess}
\title{GSEA pre-processing step}
\usage{
arseq.gsea.preprocess(deg)
}
\arguments{
\item{deg}{Differentially expressed genes dataframe returned by DESeq2 analysis.}
}
\value{
ranked list of genes.
}
\description{
Function takes in the differentially expressed genes from DESEq2 and produces a ranked list for GSEA.
}
\examples{
ranked.list <- arseq.gsea.preprocess (deg=example_deg)
}
