% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseq.gsea.R
\name{arseq.gsea}
\alias{arseq.gsea}
\title{Gene Set Enrichment Analysis (GSEA)}
\usage{
arseq.gsea(
  ranked.list,
  custom.gsea = NULL,
  f.name = "GeneSet",
  fsub.name = "enrichment",
  category = NULL,
  subcategory = NULL,
  save = FALSE,
  save.dir = getwd()
)
}
\arguments{
\item{ranked.list}{Ranked list of genes for GSEA analysis. Use 'arseq.gsea.preprocess' to generate ranked list of genes from DESeq2 differential gene expression analysis.}

\item{custom.gsea}{User defined gene list to perform GSEA. File need to be supplied as a dataframe with each row as a gene list}

\item{f.name}{Folder name to save the results when 'save=TRUE'}

\item{fsub.name}{File name of the results when 'save=TRUE'}

\item{category}{collection, such as H, C1, C2, C3, C4, C5, C6, C7}

\item{subcategory}{sub-collection, such as CGP, MIR, BP, etc}

\item{save}{Logical. Argument to indicate if to save the results in disk (pdf and csv files). Default=FALSE}

\item{save.dir}{Directory to save the results when 'save=TRUE'. Default= Working Directory.}
}
\value{
List containing a dataframe and a ggplot object.
}
\description{
Performing GSEA enrichment analysis on a ranked list of genes.
}
\examples{
ranked.list <- arseq.gsea.preprocess (deg=example_deg)
gsea.output <- arseq.gsea (ranked.list, category = "H")
}
