% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseq.kegg.enrich.plot.R
\name{arseq.kegg.enrich.plot}
\alias{arseq.kegg.enrich.plot}
\title{Gene Ontology (GO) term enrichment}
\usage{
arseq.kegg.enrich.plot(kegg.enrich, foldchanges, save.dir = getwd(),
  pathway.plots = TRUE)
}
\arguments{
\item{kegg.enrich}{Enrichment dataframe. Output of the function 'arseq.kegg.enrich'}

\item{foldchanges}{Foldchange information to overlay on the pathway plots.}

\item{save.dir}{User defined directory to save the plots in. Default is working directory.}

\item{pathway.plots}{Logical.Parameter to indicate if the pathways figures need to be drawn.}
}
\value{
GO enrichment plot
}
\description{
Plotting the top 5 enriched term from each category
}
\examples{
\dontrun{
kegg.enrich <- arseq.kegg.enrich (example_deg)
kegg.plot <- arseq.kegg.enrich.plot (kegg.enrich, foldchanges=kegg.enrich$foldchanges)
}
}
