% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseq.kegg.enrich.R
\name{arseq.kegg.enrich}
\alias{arseq.kegg.enrich}
\title{KEGG pathway enrichment}
\usage{
arseq.kegg.enrich(deg, kegg.compare = "as.group")
}
\arguments{
\item{deg}{Differentially expressed genes dataframe returned by DESeq2 analysis}

\item{kegg.compare}{Character, which comparison scheme to be used: 'paired', 'unpaired', '1ongroup', 'as.group'. 'paired' is the default, ref and samp are of equal length and one-on-one paired by the original experimental design; 'as.group', group-on-group comparison between ref and samp; 'unpaired' (used to be '1on1'), one-on-one comparison between all possible ref and samp combinations, although the original experimental design may not be one-on-one paired; '1ongroup', comparison between one samp column at a time vs the average of all ref columns.}
}
\value{
Enrichment dataframe
}
\description{
Performing KEGG pathway enrichment analysis on the significantly differentially expressed genes.
}
\examples{
\dontrun{
kegg.enrich <- arseq.kegg.enrich (example_deg)
}
}
