% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseq.kmeans.plot.R
\name{arseq.kmeans.plot}
\alias{arseq.kmeans.plot}
\title{K Means clustering Heatmap}
\usage{
arseq.kmeans.plot(
  data,
  clusters,
  metadata,
  intgroup = "arseq.group",
  save.plot = FALSE,
  save.dir = getwd()
)
}
\arguments{
\item{data}{Normalized expression dataframe}

\item{clusters}{List of gene clusters returned by Kmeans clustering algorithm}
}
\value{
Heatmap
}
\description{
Heatmap of the Most Variable genes divided into clusters
}
\examples{
arseq.kmeans.plot (data,clusters=kmeansclusters,metadata= example_meta,intgroup="treatment")
}
