% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseq.kmeans.R
\name{arseq.kmeans}
\alias{arseq.kmeans}
\title{K Means clustering}
\usage{
arseq.kmeans(data, kmeans = kmeans)
}
\arguments{
\item{data}{Normalized expression dataframe}

\item{kmeans}{integer: The number of clusters to be returned}
}
\value{
Kmeans clusters
}
\description{
K means clustering to identify gene modules
}
\examples{
clusters <- arseq.kmeans (data, kmeans=10)
}
