% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseq.kmeans.reactome.plot.R
\name{arseq.kmeans.reactome.plot}
\alias{arseq.kmeans.reactome.plot}
\title{Reactome Pathway Enrichment of Kmeans clustering output}
\usage{
arseq.kmeans.reactome.plot(clusters, save.plot = FALSE, save.dir = getwd())
}
\arguments{
\item{clusters}{List of gene clusters returned by Kmeans clustering algorithm}
}
\value{
DotPlot of Reactome Enrichment
}
\description{
Reactome Pathway Enrichment of Kmeans clustering output
}
\examples{
arseq.kmeans.reactome.plot (clusters)
}
