% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseq.mds.plot.R
\name{arseq.mds.plot}
\alias{arseq.mds.plot}
\title{Multidimensional scaling (MDS) Analysis}
\usage{
arseq.mds.plot(dds, intgroup = "arseq.group")
}
\arguments{
\item{dds}{DESeq2 object}

\item{intgroup}{interesting groups: a character vector of names in colData(x) to use for grouping.}
}
\value{
ggplot of the two selected principal components
}
\description{
Plots the Multidimensional scaling (MDS) factors.
}
\examples{
\dontrun{
mds.plot <- arseq.mds.plot (example_dds, intgroup="treatment")
}
}
