% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseq.mvg.plot.R
\name{arseq.mvg.plot}
\alias{arseq.mvg.plot}
\title{Most Variable Genes}
\usage{
arseq.mvg.plot(mvg, metadata, intgroup = "arseq.group",
  save.plot = FALSE, save.dir = getwd())
}
\arguments{
\item{mvg}{Expression matrix of the most variable genes identified using the function 'arseq.mvg'}

\item{metadata}{A CSV file with information regarding the samples.}

\item{intgroup}{interesting groups: a character vector of names in colData(x) to use for grouping.}

\item{save.plot}{Logical. Argument to state if the plot needs to be saved in disk. Default: FALSE}

\item{save.dir}{Location to save the plot when 'save.plot=TRUE'. Default: Working Directory.}
}
\value{
Heatmap of the most variable genes
}
\description{
Identifying the most variable genes in the dataset
}
\examples{
mvg <- arseq.mvg (example_dds)
mvg.plot <- arseq.mvg.plot (mvg[1:100,],metadata= example_meta,intgroup="treatment")
}
