% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseq.mvg.R
\name{arseq.mvg}
\alias{arseq.mvg}
\title{Most Variable Genes}
\usage{
arseq.mvg(data, variable.genes = 1000, dds.object = TRUE)
}
\arguments{
\item{data}{DESeq2 object or normalized expression dataframe}

\item{variable.genes}{numeric: The number of most variable genes to be identified. By default, the program identifies the top 1000 most variable genes.}

\item{dds.object}{Logical parameter indicating if the data is a DESeq2 object. Default=TRUE}
}
\value{
The most variable genes
}
\description{
Identifying the most variable genes in the dataset
}
\examples{
mvg <- arseq.mvg (example_dds)
}
