% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseq.pca.plot.R
\name{arseq.pca.plot}
\alias{arseq.pca.plot}
\title{Principal Component Analysis}
\usage{
arseq.pca.plot(
  dds,
  intgroup = "arseq.group",
  ntop = 500,
  pc.a = 1,
  pc.b = 2,
  returnData = FALSE
)
}
\arguments{
\item{dds}{DESeq2 object}

\item{intgroup}{interesting groups: a character vector of names in colData(x) to use for grouping.}

\item{ntop}{number of top genes to use for principal components, selected by highest row variance}

\item{pc.a}{numeric: Defines the first principal component. Default:1}

\item{pc.b}{numeric: Defines the second principal component.Default:2}

\item{returnData}{should the function only return the data.frame of 'pc.a' and 'pc.b' with intgroup covariates for custom plotting (default is FALSE)}
}
\value{
ggplot of the two selected principal components
}
\description{
Plots any two specified principal components. By default it plots the first two principal components.
}
\examples{
\dontrun{
pca.plot <- arseq.pca.plot (example_dds, intgroup="treatment")
}
}
