% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseq.pca.R
\name{arseq.pca}
\alias{arseq.pca}
\title{Principal Component Analysis}
\usage{
arseq.pca(dds, ntop = 500)
}
\arguments{
\item{dds}{DESeq2 object}

\item{ntop}{number of top genes to use for principal components, selected by highest row variance}
}
\value{
Eigen vectors of the Principal components
}
\description{
Plots any two specified principal components. By default it plots the first two principal components.
}
\examples{
\dontrun{
pca.ev <- arseq.pca (example_dds)
}
}
