% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseq.pheatmap.R
\name{arseq.pheatmap}
\alias{arseq.pheatmap}
\title{Pheatmap plotting function}
\usage{
arseq.pheatmap(x, filename, width = 7, height = 7, save.dir = getwd())
}
\arguments{
\item{x}{Expression matrix or any numeric matrix to be visualized as a heatmap.}

\item{filename}{The pdf file name.}

\item{width}{Width of the pdf file.}

\item{height}{Height of the pdf file.}

\item{save.dir}{Location to save the heatmap}
}
\value{
Does not return anything. The function saves a pdf file to the working directory.
}
\description{
The function saves the pheatmap as a pdf.
}
