% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseq.plot.save.R
\name{arseq.plot.save}
\alias{arseq.plot.save}
\title{Save plots}
\usage{
arseq.plot.save(arseq.plot, filename, width = 7, height = 7,
  save.dir = getwd())
}
\arguments{
\item{arseq.plot}{The plot to be saved}

\item{filename}{The pdf file name.}

\item{width}{Width of the pdf file.}

\item{height}{Height of the pdf file.}

\item{save.dir}{Location to save the plot}
}
\value{
Does not return an object. Function saves plots in directory
}
\description{
Function to save plots as a pdf file in user defined directory
}
