% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseq.poisd.dist.plot.R
\name{arseq.poisd.dist.plot}
\alias{arseq.poisd.dist.plot}
\title{Poisson distance heatmap}
\usage{
arseq.poisd.dist.plot(poisd.dist)
}
\arguments{
\item{poisd.dist}{Poisson distance matrix.}
}
\value{
Poisson distance heatmap.
}
\description{
Plots a heatmap of the Poisson distance between samples using pheatmap.
}
\examples{
poisd.dist <- arseq.poisd.dist (example_dds)
poisd.plot <- arseq.poisd.dist.plot (poisd.dist)
}
