% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseq.poisd.dist.R
\name{arseq.poisd.dist}
\alias{arseq.poisd.dist}
\title{Poisson distance}
\usage{
arseq.poisd.dist(dds)
}
\arguments{
\item{dds}{DESeq2 object}
}
\value{
Poisson distance matrix
}
\description{
Calculate the Poisson distance between samples.
}
\examples{
poisd.dist <- arseq.poisd.dist (example_dds)
}
