% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseq.reactome.enrich.R
\name{arseq.reactome.enrich}
\alias{arseq.reactome.enrich}
\title{Reactome enrichment Analysis}
\usage{
arseq.reactome.enrich(
  deg,
  Padj = 0.05,
  plot = FALSE,
  save.plot = FALSE,
  save.dir = getwd()
)
}
\arguments{
\item{deg}{Differentially expressed genes dataframe returned by DESeq2 analysis}

\item{Padj}{Numeric. Adjusted P Value cut off to define the differentially expressed genes. Default = 0.05}
}
\value{
Enrichment dataframe
}
\description{
Performing reactome enrichment analysis on the significantly differentially expressed genes.
}
\examples{
\dontrun{
reactome.enrich <- arseq.go.enrich (example_deg)
}
}
