% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseq.volcano.plot.R
\name{arseq.volcano.plot}
\alias{arseq.volcano.plot}
\title{Volcano Plot}
\usage{
arseq.volcano.plot(deg, pCutoff = 0.01)
}
\arguments{
\item{deg}{Differentially expressed genes returned by DESeq2 analysis}

\item{pCutoff}{Cut-off for statistical significance. A horizontal line will be drawn at -log10(pCutoff). DEFAULT = 0.01. OPTIONAL.}
}
\value{
Volcano plot
}
\description{
Volcano plot of the differentially expressed genes.
}
\examples{
\dontrun{
volcano.plot <- arseq.volcano.plot (example_deg)
}
}
