% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{.onAttach}
\alias{.onAttach}
\title{Startup Message}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{Library Name}

\item{pkgname}{Package name}
}
\description{
ARSeq- An automated RNASeq analysis pipeline
}
