% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_deg.R
\docType{data}
\name{example_deg}
\alias{example_deg}
\title{Differentially expressed genes}
\format{dataframe}
\usage{
example_deg
}
\description{
Example Differentially expressed genes dataframe
}
\keyword{datasets}
