% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_meta.R
\docType{data}
\name{example_meta}
\alias{example_meta}
\title{Meta data}
\format{
dataframe
}
\usage{
example_meta
}
\description{
Example meta data file. Follow this structure to make your own meta data file. Please note that the order of the samples and their names need to be exactly same as that of the counts table.
}
