% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseq.ensembl2genename.R
\name{arseq.ensembl2genename}
\alias{arseq.ensembl2genename}
\title{Gene ID converter}
\usage{
arseq.ensembl2genename(data)
}
\arguments{
\item{data}{Expression matrix with ensemble id's as the first column}
}
\value{
Expression matrix with gene names as the first column
}
\description{
Convert ensembl id's to hugo gene names. The function also reduces multiple transcripts into a single gene by taking the sum of all transcripts.
}
\examples{
\dontrun{
contrast = list(A = c("control"), B= c("drug_A"))
arseq (data = example_data,meta = example_meta, design = "treatment", contrast = contrast)
}
}
