% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arseq.R
\name{arseq}
\alias{arseq}
\title{Automated RNASeq Analysis Pipeline}
\usage{
arseq(data, meta, design, contrast, qc = TRUE, dgea = TRUE,
  variable.genes = 1000, folder.name = "ARSeq", custom.gsea = NULL,
  save.dir = getwd())
}
\arguments{
\item{data}{Un-normalized counts matrix (please note that you should NOT pass in normalized data).  The counts' table should contain unique gene names as the first column. ENSEMBL ID's are also allowed but no other form of ID's are currently supported. Check example- head(example_data): \code{\link{example_data}}.}

\item{meta}{A CSV file with information regarding the samples. It is absolutely critical that the columns of the counts' matrix and the rows of the metadata are in the same order. The function will not make guesses as to which column of the count matrix belongs to which row of the metadata, these must be provided in a consistent order. Check example- head(example_meta): \code{\link{example_meta}}.}

\item{design}{The formula that expresses how the counts for each gene depend on your metadata (used to calculate the necessary data for differential gene expression analysis). Check DESeq2 documentation for designing the formula. In general, you pass in a column name (e.g. treatment) of your metadata file or a combination of column names (e.g. treatment + cell_type).}

\item{contrast}{Information regarding the groups between which you would like to perform differential gene expression analysis. It could be between two groups or between multiple groups and needs to follow the following format: contrast = list(A = c(" "), B= c(" ")). If you are comparing two groups (e.g. control vs treatment), the constrast argument should look like the following: contrast = list(A = c("control"), B= c("treatment")). In situations where you have multiple groups to compare- (e.g. control vs treatment1 and treatment2), you should do the following- contrast = list(A = c("control"), B= c("treatment1", "treatment2")).}

\item{qc}{Logical. When passed in TRUE, the program would run the quality control modules on the entire dataset. If you are planning to perform multiple comparisons using the contrast argument, run  qc = TRUE for the first time and then change it to qc = FALSE for the subsequent comparisons to speed up the analysis.}

\item{dgea}{Logical. Parameter to define if differential gene expression analysis is to be performed. Default: TRUE}

\item{variable.genes}{numeric: The number of most variable genes to be identified. By default, the program identifies the top 1000 most variable genes.}

\item{folder.name}{Custom folder name that you would like to save your results in.}

\item{custom.gsea}{User defined gene list to perform GSEA. File need to be supplied as a dataframe with each row as a gene list}

\item{save.dir}{Directory to save the results in. Default: Working Directory.}
}
\value{
Differentially expressed genes.
}
\description{
An easy to use pipeline for analysing RNA Seq data. The package currently supports the following analysis- Differential gene expression analysis using DESeq2, Calculate the most variable genes, PCA analysis, GO enrichment of the differentially expressed genes, KEGG pathway enrichment of the differentially expressed genes, GSEA analysis.
}
\examples{
\dontrun{
contrast = list(A = c("control"), B= c("drug_A"))
arseq2 (data = example_data,meta = example_meta, design = "treatment", contrast = contrast)
}
}
