% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R, R/example_meta.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Counts table}
\format{dataframe}
\usage{
example_data

example_meta
}
\description{
Example counts table.

Example meta data file.
}
\keyword{datasets}
