% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R, R/example_meta.R,
%   R/kegg.sets.hs.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Counts table}
\format{dataframe}
\usage{
example_data

example_meta

kegg.sets.hs
}
\description{
Example counts table.

Example meta data file.

kegg.sets.hs
}
\keyword{datasets}
